HELP.md updated
HELP.md updated

README.md updated

README.md updated

+++ src/main/java/com/example/personal_notes_app/Model/Articles.java
@@ -13,6 +13,6 @@ public class Articles {
     private String articleTitle;
     private String article;
     @Id
-    private int articleID;
+    private long articleID;
 
 }
